requirements.txt
Django==4.2.*
djangorestframework==3.15.*
psycopg2-binary
python-dotenv
python-decouple
dj-database-url
